Getting it all to work together.
Django is probably not going to be useful as an API for MongoDB.
It also seems not to be useful for severing static files. Django might be
removed in favor of a Pure python docker container that is used to populate
the MongoDB database.

The TeleView app in next.js is working and able to display from a docker image.
It is able to read from the MongoDB database directly, bipassing the need for
an API. However, it is still displaying the default page, and has no real
functionality build in.

I am continuing to struggle to get NGINX to serve static files needed to
populate the web portal's views with images and data. This is frustrating
because it should be simple, but perhaps it is not allowed from a docker
volume? Research is ongoing into this matter.