Using server side async functions for Query of MongoDB
This will allow use to dynamically populate Home, and other pages, with data
from the MongoDB database.

Server side rendering means that all the information is gathered before sending
the HTML and image to the client's browser. This should be the fastest render
possible for distant remote connections.

A refactor to organize the style files (CSS) in a styles directory,
favicons, and other public data in the public directory that will also contain
all other public assets. The app will only contain files that look like the
URI layout of the website, inline with what is recommended for Next.js in 2023.