Separated mongoDB into its own compose file.
We need the mongoDB container to be up and running before the build the next.js
production build. This seems to be a larger and still open problem.
The best solution for now is to separate the `docker compose up` commands into
a cascading number of commands that can be called be a single shell script.
This is not an ideal solution as it adds one more layer of complexity above
docker compose, but this does seem to be the recommended (and least complex)
solution.