Added Bootstrap and Jquery to the static files.
Everything was working before and is working now. This change is to stop the
Dango-API from requiring the internet to retrieve static asset like
Bootstrap and jQuery. These are now locally installed and will ship with
the TeleView repository.

This is a necessary upgrade to have TeleView continue to work
"off-the-grid" in remote locations with, or without consistent internet access.