Big updates to Django views.
Added a framework to set status for database actions. For now, this simply
uses the build-in sqlite database that is Django's default, but this could
be migrated to mongo or else where if needed.

Data processing statuses can be set via a GET request that passes the required
information in a special formatted URL. See the example below
f'/api/post_status/{status_type}={percent_complete}'

Data statuses can be read via a JSON response at the URL /api/get_status

Django only generates webpages on the server side; however, it is possible to
run client side code on written in JavaScript. A framework for doing this is
being used to have JavaScript code to check the completion status of database
actions and refresh the page content until all processes are complete.

Test programs show that this is working in the development environment
and in production.

A major projection upgrade was required to serve static files, which includes
JavaScript, images, and CSS files. The Django part of the website is using
Bootstrap to make the CSS changes a little simpler for a website that will not
be edited frequently. Django really only and API between the web and python
database processes.

The next step is to upgrade the Python database scrapping code to interact with
the new web-based interface.