Production Website triggers database scanning actions
There is a lot in this update that contributed to a working *Production* method
for triggering database actions in python.

The large bug was that a request cannot be made on the same thread that is
handling the initializing request. This rule makes sense, and there was a
better way to update the database direction through the StatusModel class.
This was only causing issues in the production build for the website.

A reporting system was designed to report data scanning status to a URL
that updates the local SQLite database. This database is used to store status
information that triggers status loading messages and refresh commands on the
api/ main page. The reporting system is designed not over a report. This makes
the reporting a bit slow compared to the scanning process for very small
databases.

This commit introduces a new file type `.env`. This is a file of configuration
options used in the compose.yaml file. Git ignores this file,
to see how file this could be used to configure a give system see .env-default,
a new file in this commit.

By default, an admin page for Django at /api/admin/. This is a good tool for
debugging processes and resetting database values. This user is set at the
image build time with optional arguments that must all be supplied to for
success. When using docker compose, these values are set with the .env file.