Docker compose .env variables made available to NestJS app
The major upgrades for getting the files into the NextJS app which doesn't use
environment variables from unix, but instead insists that environment variables
be specified in a file. This file is called .env.production when we consider
only the production case of creating a docker image of the NextJS app.

1. These variables read from the .evn file and are propagated through
docker compose.yaml to the dockerfile at build-time as compose-ARGS.

2. In the docker file, the compose-ARGs are added as
temporary environment variables using the Docker ARG command.

3. These environment varaibles are now available in docker image that is
being constructed. We then use the RUN with unix `echo` command to write a
file called the .env.production (see .env.production.example). This echo
command has a string with environment variables values that will be evaluated
when the echo command runs, as in

`Variable_In_File=$ENV_VARIABLES`.
`>>>Variable_In_File=MY_CUSTOM_VALUE`

In this process, we create a file just-in-time file based on variables specified
in the global project file, .env.

I have also switched to the "key: value" method for specifying environment
and argument variables in compose.yaml. This syntax is "nudge" that tells a
reader that these parameters must have unique keys that should not be
specified twice.

Added a .dockerignore file to exclude .evn.production file since that file is
made dynamically at build-time for tvapp/Dockerfile.