New default page that will be used as a filter form.
A new default page (Homepage) has been deployed. This is a work in progress,
but important elements have been added that allow for a very large namespace
to be used for generating form URL that will allow filter views to be potable
for URL based sharing, but also for doing database queries only on the
sever-side rendering.

A very large name spaces available for the default filter form page;
this which allows the form to expand in complexity and scope over the lifetime
of the project.

The existing process for data access has been left intact and operational.
This serves as a minimum operational structure to view data and check systems
while the more general filter form is being designed. Eventually, these views
will be deprecated and removed from the codebase.

Quality-of-life improvements:

- Added custom colors to the tailwind.config.js.

- Learned how to use Tailwind CSS to make the rows and columns needed to present
the default filter form pages to as a single page dashboard style application.

- Spelling fix: time_stamp -> timestamp

- Spelling fix: course -> coarse

- MongoDB smurf-database has a new field for its documents, umf_label.
This is a non-indexed combination of ufm_letter and ufm_number in the expected
format: `f'{ufm_letter.upper()}v{ufm_number}'`.