Table added for navigating to filtered items.
The Primary upgrade is a table showing the database items with links to
that item's plot and data files. This table is a major component in how users
will navigate to their data. Any combination of database filters will display
results in the new table.

When table results are over a certain size (100 is default), only a limited
number of items is displayed. Limiting the number of items returned should help
to prevent over using the server's resources for more data, particularly
because that volume of data is not useful to our users. Limiting the number of
items means that we need to provide users with a way to navigate to other
groups of items.

The table has 4 navigation elements for browsing First, Prev, Next, Last. These
navigation tools work as expected to browse filter results that have large
numbers of items. However, the data is sent from the app-server each time a
navigation element is clicked. The navigation elements simply redirect to a
pages where the requested state is available.

It is good to have these permanent urls that have has some of the website's
state variables encoded. This allows sharing data views by simply sharing URLs.

The URL encoding for navigation is more general than what is being used in the
current navigation elements. In the future, the navigation can be upgraded with
form elements that allow users to write and select the ranges they want to view.

The styling and viewing of the data is also processed, but it is far from a
finished product.