Bugs fixed: List
1. Non unique key names for navigation elements have been updated to be unique.

2. All internal URLs in the app are now prepared with a "/". This prevents the
use of an empty string URL "", which apparently is reserved for a redirect to
the current page being viewed.

3. I disabled prefetch for all the links. Hovering over a link was causing
the link page to be requested and stored in the browser's cache. This made
navigation seem instant. However, I felt concerned about loading 20 pages in
the background while only a single page is being viewed. This can be
reactivated at any time, and for specific links.

4. Updated the mongoDB connection. It was working fine in production mode,
however, in development it was adding and dropping too many connections to
the MongoDB database. The database connection is now stored as a
global variable in the development environment only that persists across
reloads. The client connection to MongoDB was moved to its own file,
client.tsx