Excluding `None` timestamp folders from a database.
This resolves issue #4. None folders are excluded from the database.
These excluded directories are written to MongoDB in a `logs` database.
These errors are reported as they are encountered, but only unique errors,
enforced by a unique combination of (dir_type, dir_path, reason) entries, are
retained in the database. However, the `error_last_seen_timestamp` is updated
everytime an issue is observed.

This issue was addressed since the typing of the timestamp variable when it was
allowed to be `None` was causing the code to be complex with no benefit once
this issue was resolved.

This is the beginning of a solution for #1, but it is not completely implemented.
Thus, it will not be marked as resolved until a subsequent commit.