Added a timestamp range filter.
This is an intermediate commit. The projects is not filtering the database
with timestamp ranges yet.

It is updating the links and state variables for the search page that contain
the timestamp range information. The mock version of these new state variables
is tested and is working as expected.

This updates includes the first client-side reactive component. This is in
the form of input text boxes that convert timestamps to date and visa-versa.

There was a major refactor to make the types using for data filtering to be
strict and explicit. This added a lot of (very readable) boilerplate code.
This is a bit of an experiment in embracing typescript.

Previous data views have been removed as they are no longer used/supported.

There are some bugs, a stable build will follow with some other quality-of life
improvements for working with timestamp ranges.