First full implementation of timestamp range filter
1. Over lapping timestamp ranges are automatically combined on read-in and
generated URL representations. This is important for only getting unique
document sets when querying the database.

2. Bug fix for data_view URL that needed a new pipeline to access the
database cursor after there were changes to the cursor function to make typing
much more restricted.

3. Bug fix for modifierState types. The document range function was not working
after the URL parsing refactor for stricter filterState typing.

4. Fixed an issue with the production build. I had a function name
that was not one of the three allowed names in a Page.tsx file. This was
causing the docker build to fail. This function was moved to an appropriate
location. page.tsx files should be the JSX Page() and all other components
must be in other directories.

5. The process of creating the filter object that provides instructions to the
MongoDB find() operation needed to be upgraded to handle ranges. This upgrade
was accomplished while also making an improved abstract framework for
generating complex filters. The filter object is also writer to generate
a less complex, and more readable, object when queries are single values
and not based on arrays of items.

6. The TELEVIEW_VERBOSE boolean now toggles all console.log() statements.
It is now possible to run silently based on this variable that is settable
from docker compose.