Added state-context for sharing state data across elements.
We are now able to toggle the display for a crude menu for removing
filter components from the database query.

The display state (true or false) of that menu is the first variable that is
stored in the client-side components that track what view-state the user has
selected based on button clicks and other interactions.