Move the Remove filter links to a Drawer
This is a button triggered Drawer that can be reused to display results for
many menus. The new layout is starting to take shape with recognized elements
and navigation buttons on the top of the page.

Database Queries are now preformed on a server-side database.
The data is then transferred to the client-computer rendered functioned that
will render the page in the context of user-selected states. These states will
be driven by the user interactions with the data already received.