Refined Drawer component, extended it for general use
This abstraction will allow us to rapidly make new drawers for any new
menus that are developed.

A lot of code from the original example was not needed for this application.
For example, there is no reason to access and control the DOM attributes
directly. The Drawer component is now simple and tailored for our usage.

Drawers are now additional layers built to display over the top of the main
viewing area. The top most layer is the menu for users to select from several
options and includes a close button. The middle layer uses a blurring effect
to obscure the main content, thus bring focus to the menu in the drawer.
Clicking on anywhere on the middle layer also closes the menu and restores the
main view and resets the state variables. These layers are managed with the
z-index attribute in Tailwind CSS that creates a z-index HTML component.