Upgraded all existing components for the Layout Roadmap
Buttons are styled and work well.

The code is simplified with a clearly defined boundaries between the server and
the client.

The data context is now hydrating all the rendered states directly, as
opposed to passing argument when calling functions.

Still remaining is to build a single data viewer on the black half of the
webpage. However...

We are going to push these stable upgrades to the main branch and pivot to
Getting ready for the first production installation.