Runtime variable management added for NEXT.js
The NEXT.js framework does not allow for environment variables except what is
written to a .env.production file. This file has been now added as
a docker shared volume. This system allows use to configure host-specific
variables at container runtime not only at build time.

The top level scripts init.sh manage the creation of this file from the existing
.env file all hosts are required to have, this file already is used to populate
run time variables for Django and MongoDB components. If no changes
are required, init.sh is only required to run a single time on a given host
computer as part of an initial setup.

Incrementing to version 0.1.3.