Serving teleview from a host with the "teleview" base-path
Teleview is now severed from somehost.com/teleview, where somehost.com
is specified in the .evn file with the variable TELEVIEW_PUBLIC_SITE_HOST.
See .evn-default for an example.

Fixed The Django "DEBUG" to convert the string environment variables to a bool.
It was always being set to True, since it was a non-empty string.

Move the tvapp favicon to a location that supports a variable base-path

The Docker named volumes were persisting after docker compose down was called.
The "-v" flag was added to the scripts automatically delete these volumes when
updating the container. This was making the Django static files never update
while the volumes existed.

Incrementing to version 0.1.4.