Added redirect for default page.
A bug in the NEXT.jS framework interferes with the use of both
a basePath (hostname.com/basPath) and the use of a default page for
dynamically generated routing. The solution was to simply redirect
the default page to a URL that provides the same information.

This should be the last fix for TeleView to work "as expected" on the
simons observatory website.