Added the "teleview" to the path for the science data products.
"teleview" being missing from the path for the science data products made these
routs to be reachable on site.simonsobs.org.

Added a line to show the parsed host name in tvapi/tvapi/settings.py in when
the DEBUG environment variable is True. This is to help debug the CORS
Permissions for the django-site.

Incremented the version number to 0.1.5