Bug for basePath not being appended in default view fixed.
This was a mistake what a leading "/" was missing from the "data_view" links in
tvapp/components/NavDocs/table.tsx. This only caused an issue when using a default page
(empty string URL) and a basePath.

Resetting all redirects and folder renames that were changed as this issue was
being investigated.

Incrementing to version 0.1.6