Many small updates, including filter by stream_id.
StreamID, an extremely repetitive component for the smurf data directory structure,
is not natively supported for database indexing and filter functions. Stream ID
is formatted as ufm_{a_letter}v{a_number} where a_number is the only unique piece
of information.

Stream ID was added as a data column; however, it was hard enough to read
that I left the complimentary UFM Label in place.

Stream ID was added as an option to for matching filters, allowing users to
select only the UFMs of interest. Since Stream ID is an indexed values in the
database, I don't expect any performance degridation compared to simply
filter by the UFM number, the only unique information in the Stream ID.

The table is now sorted in reverse chronological order by default. This means
that the newest data will always be at the top for any given filter state.
Anticipating users can simply refresh the page at any filter state to get the
latest values for a given filter state.

A human-readable DataTime, generated from the timestamp value, was added to
as a column in table.tsx.

Incremented to version 0.1.7