Added a data descriptor for "platform" (telescope).
This is a change that touches all levels of this project.  The platform
descriptor is parsed from the top level of a folder hierarchy.
The next level of the folder hierarchy is data type, with "smurf" being
only configured data type to parse.

This is a generalization of with a previous data framework that pointed to
a single directory called smurf. Now a directory with N unique platform
names, each of with is capable of having its own "smurf" data directory.

The change to include "platform" as a data descriptor was also propagated
to the front-end and is now included in the displayed data and data
filtering options.

Better descriptions were added to the Django database interface website.
This includes an acknowledgment that what is being triggered is a
"Full database reset" but what is tracked can actually several
individual processes.