New Feature - Job Scheduler
The job scheduler is a custom system that only allows one job of the
same name in the job queue. When a job is finished running, the next job is
popped from the queue and is called to run. These asynchronous states are stored on
a local SQLite database (in the docker container, or the localhost in development).

These states are viewable on the tvapi default page. There are also many links
that clear/set states and add jobs to the queue.

The Scheduler jobs are incrmented by an event-loop that is called with a cadence
of greater than one second; this gives the database time to update that a job
is running and to then lock the queue. The interval is controlled in settings.py
with the variable SCHEDULER_SLEEP_TIME_SECONDS.

This process is simulated once a minute by a cron-job that calls a specific
database view via http request. This was done as the cron environment does not
easily interact with the environment variables required for
database interactions. Worse, the cron does a fresh import every time.

There are many views that are available as links on the default pade for
tvapi. These links are helpful for testing and developing the scheduler.
However, a reduction and simplification of the available processes is
needed for effective user communication.