simplified query status variables and states
All task and scheduler states are now set using the "running" var_name value
in the table model SchedulerState.

Added a hook to retry failed actions by adding them back into the queue.
The status 'failed' also allows the queue to advance.

Added a new tool category for that for the tvapi default page that only
shows up when DEBUG is set to True in settings.py.