Added a uniqueness constraint to smurf documents.
Adding duplicate data is now impossible as a unique index is required that mimics
the constraints of the 'smurf' directory system. This solves GitHub Issue #1.

A new database action, 'update', was added. This allows a files system scan to
non-destructively update an existing collection. This is a tool that can be used
without causing a noticeable disruption to the tvapp-website while still updating
mongoDB with changes from the files system.

Incremented version number to 0.1.10