Views added for updating mongoDB periodically.
These views add functionality to update only parts of the database.
One view is finished, it updates all folders with a timestamp that is
within 300,000 seconds of the current time. This is meant to run as
a crontab once per day. When running continuously, this will add a
safety net to ensure that the mongoDB database is in-sync with the
static file system. This addresses and solves GitHub issue #3.
A crontab was added to automatically queue this database action
once-per-day in the deployed system.

A second view was added that is currently incomplete. This
view is meant to update the mongoDB database every five minutes.
To do this, we are starting to keep track of the maximum last
modified time for the file system. This will allow only the
selection of new files for an update for this high-cadence update.
A state-tracking system for last-modified-folder-time is deployed in this commit.
This is required to address GitHub issue #2.

Many internal database tools were added to manage the new state variables.
Include the new file database_states.py.