5-minute database update, GitHub issue #2.
A short-term database update is now an available feature. On the production
website, this feature is called every 5 minutes via a crontab in specified
in settings.py. This tool is designed to only upload data to MongoDB when
the modification time of the directories with data files is more recent
than the last database scan. This is directly addressing GitHub issue #2.

An additional feature upgrade is that the crontab database actions will now
also initialize an empty database and also make long updates in the event
that the file system continues to accumulate data while the website is not
running. Ideally, users will never need to use the database interaction
tools on the Django tvapi.

Since the update cadence is 5 minutes, and updates are no longer a
destructive process, the link the Django tvapi link-button is now removed
from the tvapp default page for the prodcution version of the main website.
Data is designed to simply appear in the database used by the front-end
website, tvapp, every 5 minutes without any user interaction.

The Django API default page has been revised and given a high level
of attention. In lue of documentation on a separate page, instructions
are with as div-titles (hover-text). Verbiage and capitalization have
standardized. This is likely the last major update for the view and look
of the tvapi default page. Only small, and self-similar updates are on
the horizon.

This marks the end of the alpha-development-phase for the TeleView backend.