New data view for the front-end
Enabled rending of server and client side data-views.

Added the data-view as half of the teleview default page.

Modularized table columns to be added and removed quickly.
This could lead to an upgrade to allow users to add and remove columns
at will.

Removed async from the image loading tool, which was crashing the
client-side rendering in an infinite loop.

Changed data_view to be smurf_data_view, in the hope that this website might
one day serve a different data system.

Updated all the unique document data identifiers to be exactly
what is enforced in MongoDB as the unique compound index for
smurf documents. Ensured that the unique compound index components
are in an order that reflects the file system data that this index
is originally parsed from.

Added a new variable to the shared component context for tracking the
currently selected data-view from the table. Subsequent updates will
do more with setting and using this to Navigate documents.

Incrementing to version 0.1.13