Attempt to address a Queue increment failure
On August 19th, the Queue failed to increment on the site.simonsobs.org.
This left it in a 'locked' status that could never be cleared. Manually
clearing the issue on the 20th solved the issue, the issue did not reoccur.

This commit attempts to fix these issues. It assumes that this was an SQLite
OperationalError exception. This exception is now wrapped in a try statement,
that will retry to increment the 20 times, sleep for 2-seconds between tries.

I was not able to find the error in the docker logs. This was surprising,
leading me to think either the error was not logged or the thread was killed.
A killed thread seems unlikely.

However, this leads me to find that the logs were populated with VERBOSE
statements that were not expected. This error was easy to find VERBOSE
was being imported from two places, and in configs.py, the import was
incorrectly parsing the bool environment variable.
To simplify the configuration, I have deleted configs.py and moved all
environment variable parsing a distribution to setting.py, where bool
environment variables were being correctly parsed.

Incremented to version 0.1.14, passing the work on the dataview branch.