Completed Basic layout and navigation for Smurf-data-view
It is possible to navigate the smurf-data-view while using the Navigation
table. This works with:

1. Clicking a table-row
2. by pressing up, down, right, left arrow keys and enter button
3. using float arrow key buttons.

The freestanding link to this data view is still available allowing
a link that is as permanent as the MongoDB database record.

Some styling and formatting based on testing will be needed before
a pull-request to the `main` branch.