Ready to release version 0.2.0
Many small updates as a result of testing.

Added tool to copy a freestanding data-view link to the clipboard.

Add a pop-up menu primitive component. This is only used to be the
copy-data-view-to-clipboard tool, but this can be extended to other user
messages that appear temporarily. Much like the menu system is made
from re-usable components.