Updating the docs with the application's realized structure.
The docs describe a system I once intended to build.
TeleView's current system is different and much richer, the docs
are being updated to reflect those upgrades.

Two new images to description the expected data topology
and the networked application topology.

New text for getting the application started for the first time on a
new test-computer or data-host computer.

Incremented to version 2.1