Update to the JS packages used.
GitHub Dependabot found and issue, I updated all the dependencies and
tested the docker build locally it appears to be working.

Minor name change in find.py to clarify the purpose of a variable that
refers to the `platform name`. It is not just the value of a directory,
but is also a name used for sorting and indexing data.