Docker service starting order change, mongodb starts before tvapi
This addresses one of the points in GitHub issue #17.

The pythonAPI service (tvapi) will now only start once the
MongoDB service is already running and healthy.
We are not able to migrate all database components used by Django
to a MongoDB database.

The remaking task remains is to switch from SQLite to mongoDB in the
tvapi/tvapi/settings.py variable `DATABASES`